basic code is done

+++ src/main/java/lv/marmog/zone2/zone2/controllers/AuthorController.java
@@ -16,15 +16,12 @@ public class AuthorController {
     @Autowired
     private AuthorService authorService;
 
-
-    /* to get all the authors */
-
-    @GetMapping("/author")
-    public ResponseEntity<List<AuthorDTO>> getAllauthors() {
+    @GetMapping("/authors")
+    public ResponseEntity<List<AuthorDTO>> getAllAuthors() {
         try {
             List<AuthorDTO> authorsDTO = authorService.getAuthors();
 
-            if (authorsDTO.isEmpty() || authorsDTO.size() == 0) {
+            if (authorsDTO.size() == 0) {
                 return new ResponseEntity<>(HttpStatus.NO_CONTENT);
             }
 
@@ -33,9 +30,6 @@ public class AuthorController {
             return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
-
-
-    /* to get a particular author */
 
     @GetMapping("/author/{id}")
     public ResponseEntity<AuthorDTO> getAuthor(@PathVariable Integer id) {
@@ -48,48 +42,56 @@ public class AuthorController {
         return new ResponseEntity<AuthorDTO>(HttpStatus.NOT_FOUND);
     }
 
-
-    /* Controller to handle adding a new author */
-
-    @PostMapping("/author")
-    public ResponseEntity<AuthorDTO> saveAuthor(@RequestBody AuthorDTO authorDTO) {
+    @PostMapping("/author-create")
+    public ResponseEntity<AuthorDTO> createNewAuthor(@RequestBody AuthorDTO authorDTO) {
         try {
             return new ResponseEntity<AuthorDTO>(authorService.addAuthor(authorDTO), HttpStatus.CREATED);
         } catch (Exception e) {
-            e.printStackTrace();
             return new ResponseEntity<AuthorDTO>(HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
 
 
-    /* Controller to handle updation of a author */
-
     @PutMapping("/author/{authorId}")
     public ResponseEntity<AuthorDTO> updateAuthor(@RequestBody AuthorDTO authorDTO, @PathVariable Integer authorId) {
         try {
-            //get author record to be updated
+
             AuthorDTO authorToUpdate = authorService.getAuthorById(authorId);
 
-            //set the values
             authorToUpdate.setAuthorName(authorDTO.getAuthorName());
 
-            //update the changes
             return new ResponseEntity<AuthorDTO>(authorService.addAuthor(authorToUpdate), HttpStatus.OK);
         } catch (Exception e) {
             return new ResponseEntity<AuthorDTO>(HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
 
-
-    /* Controller to handle deleting a author */
-
     @DeleteMapping("/author/{id}")
     public ResponseEntity<Author> deleteAuthor(@PathVariable Integer id) {
         try {
+            AuthorDTO authorToUpdate = authorService.getAuthorById(id);
+            if(authorToUpdate==null){
+                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            }
             authorService.deleteAuthor(id);
             return new ResponseEntity<>(HttpStatus.NO_CONTENT);
         } catch (Exception e) {
             return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+    }
+
+    @GetMapping ("/authors-find-by-name")
+     public ResponseEntity <List<AuthorDTO>> getAuthorWithName (@RequestParam String name) {
+        try {
+            List<AuthorDTO> authorsDTO = authorService.getAuthorByName(name);
+
+            if (authorsDTO.size() == 0) {
+                return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            }
+
+            return new ResponseEntity<>(authorsDTO, HttpStatus.OK);
+        } catch (Exception e) {
+            return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
 }
